Add files via upload
The Shoebox is a retail online store that focuses on offering a selection of footwear. The web
application developed will rely on a database system that effectively manages information
related to the shoes sold, and the orders made by registered customers. Moreover, the database
will also encompass customer records, ensuring that only registered individuals can proceed with
placing orders.
We have been tasked with creating a database and web application for The Shoebox, an online
store specializing in the sale of shoes.